[COFF][LoongArch] Enable UseOffsetLabels for temporary symbols
For relocations against temporary symbols (that don't persist in
the object file), we normally adjust them to reference the start of
the section.

For pcalau12i relocations, the immediate offset from the referenced
symbol is stored in the opcode as the 20 bit signed immediate; this
means that the symbol referenced must be within +/- 512 KB from the
referenced symbol.

Create label symbols with regular intervals (512 KB intervals). For
relocations against temporary symbols, pick the preceding added
offset symbol and make the relocation against that instead of
against the start of the section.

This pr is inspired by the AArch64's D114340.

(cherry picked from commit c009080b4ec9938bdd2fbc5066cfecb91ada7e3a)